Add additional condition compiled file for gccgo
At least as of gccgo 10.x and up, gccgo looks for mangled C names in
//go:linkname directives. Use manually mangled name in the directive
with build constraints to make the package build with gccgo.
