+++ internal/qtls/go115.go
@@ -8,7 +8,6 @@ import (
 	"crypto/cipher"
 	"crypto/tls"
 	"net"
-	"unsafe"
 
 	qtls "github.com/marten-seemann/qtls-go1-15"
 )
@@ -97,19 +96,4 @@ type cipherSuiteTLS13 struct {
 	KeyLen int
 	AEAD   func(key, fixedNonce []byte) cipher.AEAD
 	Hash   crypto.Hash
-}
-
-//go:linkname cipherSuiteTLS13ByID github.com/marten-seemann/qtls-go1-15.cipherSuiteTLS13ByID
-func cipherSuiteTLS13ByID(id uint16) *cipherSuiteTLS13
-
-// CipherSuiteTLS13ByID gets a TLS 1.3 cipher suite.
-func CipherSuiteTLS13ByID(id uint16) *CipherSuiteTLS13 {
-	val := cipherSuiteTLS13ByID(id)
-	cs := (*cipherSuiteTLS13)(unsafe.Pointer(val))
-	return &qtls.CipherSuiteTLS13{
-		ID:     cs.ID,
-		KeyLen: cs.KeyLen,
-		AEAD:   cs.AEAD,
-		Hash:   cs.Hash,
-	}
 }
